Move edm::stream::imll::makeGlobal to own file
The function was declared in two different headers which caused problems when they were included in the same source file. This moves the functions to just one file.
